# Simplify and optimize randomElements()