Figures, on 100000 iterations:

* For several random elements: before 590 ms, after 320 ms
* For one random element: before 390 ms, after 120 ms