@vlakoff 2016-06-30
Fixed an issue with $count = 0, but remaining test failures seem unrelated…