@vlakoff 2016-06-30
Looks like it's because of the change from `mt_rand()` to `array_rand()`, which discards the `seed()`. So if you want Faker to be possibly deterministic using a seed, I guess this PR has to be dropped…